+++ src/session-client/verification-script-resource.js
@@ -1,6 +1,7 @@
 goog.module('omid.sessionClient.VerificationScriptResource');
 
 const argsChecker = goog.require('omid.common.argsChecker');
+const {AccessMode} = goog.require('omid.common.constants');
 const {packageExport} = goog.require('omid.common.exporter');
 
 /**
@@ -15,17 +16,27 @@ class VerificationScriptResource {
    * @param {string} resourceUrl
    * @param {string=} vendorKey
    * @param {string=} verificationParameters
+   * @param {AccessMode=} accessMode The level of access this verification
+   *     script will have when executed.
    * @throws error if either the vendorKey or resourceUrl is undefined, null or
    *   blank.
    */
   constructor(resourceUrl, vendorKey = undefined,
-      verificationParameters = undefined) {
+      verificationParameters = undefined, accessMode = AccessMode.FULL) {
     argsChecker.assertTruthyString('VerificationScriptResource.resourceUrl',
         resourceUrl);
 
+    /** @type {string} */
     this.resourceUrl = resourceUrl;
+
+    /** @type {(string|undefined)} */
     this.vendorKey = vendorKey;
+
+    /** @type {(string|undefined)} */
     this.verificationParameters = verificationParameters;
+
+    /** @type {!AccessMode} */
+    this.accessMode = accessMode;
   }
 
   /**
@@ -34,6 +45,7 @@ class VerificationScriptResource {
    */
   toJSON() {
     return {
+      'accessMode': this.accessMode,
       'resourceUrl': this.resourceUrl,
       'vendorKey': this.vendorKey,
       'verificationParameters': this.verificationParameters,